Removed connection-related methos from API. Moved ConnectionEvent api from store to generi connection package.
