+++ kura/org.eclipse.kura.api/META-INF/MANIFEST.MF
@@ -34,6 +34,7 @@ Export-Package: org.eclipse.kura;version="1.7.0",
  org.eclipse.kura.command;version="1.2.0",
  org.eclipse.kura.configuration;version="1.2.0",
  org.eclipse.kura.configuration.metatype;version="1.1.0",
+ org.eclipse.kura.connection.listener;version="1.0.0",
  org.eclipse.kura.container.orchestration;version="1.2.0",
  org.eclipse.kura.container.orchestration.listener;version="1.0.0",
  org.eclipse.kura.crypto;version="1.3.0",
@@ -72,7 +73,6 @@ Export-Package: org.eclipse.kura;version="1.7.0",
  org.eclipse.kura.security.tamper.detection;version="1.0.0",
  org.eclipse.kura.ssl;version="2.1.0",
  org.eclipse.kura.status;version="1.0.2",
- org.eclipse.kura.store.listener;version="1.0.0",
  org.eclipse.kura.system;version="1.6.0",
  org.eclipse.kura.type;version="1.1.0",
  org.eclipse.kura.usb;version="1.3.0",
