+++ kura/org.eclipse.kura.core/META-INF/MANIFEST.MF
@@ -32,6 +32,7 @@ Import-Package: javax.crypto,
  org.eclipse.kura;version="[1.0,2.0)",
  org.eclipse.kura.certificate;version="[2.0,3.0)",
  org.eclipse.kura.configuration;version="[1.1,2.0)",
+ org.eclipse.kura.connection.listener;version="[1.0,2.0)",
  org.eclipse.kura.core.ssl;version="1.0.0",
  org.eclipse.kura.crypto;version="[1.0,2.0)",
  org.eclipse.kura.data;version="[1.1,2.0)",
@@ -45,7 +46,6 @@ Import-Package: javax.crypto,
  org.eclipse.kura.security.keystore;version="[1.0,2.0)",
  org.eclipse.kura.ssl;version="[2.1,2.2)",
  org.eclipse.kura.status;version="[1.0,2.0)",
- org.eclipse.kura.store.listener;version="[1.0,2.0)",
  org.eclipse.kura.system;version="[1.4,2.0)",
  org.eclipse.kura.type;version="[1.1,2.0)",
  org.eclipse.kura.util.configuration;version="[1.0,2.0)",
