+++ kura/org.eclipse.kura.core/src/main/java/org/eclipse/kura/core/data/DataServiceImpl.java
@@ -43,6 +43,7 @@ import org.eclipse.kura.KuraStoreException;
 import org.eclipse.kura.KuraTooManyInflightMessagesException;
 import org.eclipse.kura.configuration.ConfigurableComponent;
 import org.eclipse.kura.configuration.ConfigurationService;
+import org.eclipse.kura.connection.listener.ConnectionListener;
 import org.eclipse.kura.core.data.store.MessageStoreState;
 import org.eclipse.kura.core.db.H2DbMessageStoreImpl;
 import org.eclipse.kura.core.internal.data.TokenBucket;
@@ -58,7 +59,6 @@ import org.eclipse.kura.message.store.provider.MessageStoreProvider;
 import org.eclipse.kura.status.CloudConnectionStatusComponent;
 import org.eclipse.kura.status.CloudConnectionStatusEnum;
 import org.eclipse.kura.status.CloudConnectionStatusService;
-import org.eclipse.kura.store.listener.ConnectionListener;
 import org.eclipse.kura.util.jdbc.ConnectionProvider;
 import org.eclipse.kura.util.jdbc.SQLFunction;
 import org.eclipse.kura.watchdog.CriticalComponent;
