+++ kura/org.eclipse.kura.db.sqlite.provider/META-INF/MANIFEST.MF
@@ -12,12 +12,12 @@ Bundle-Activator: org.eclipse.kura.internal.db.sqlite.provider.SqliteProviderAct
 Import-Package: com.zaxxer.hikari;version="2.7.9",
  org.eclipse.kura;version="[1.7,2.0)",
  org.eclipse.kura.configuration;version="[1.2,2.0)",
+ org.eclipse.kura.connection.listener;version="1.0.0",
  org.eclipse.kura.crypto;version="1.3.0",
  org.eclipse.kura.data;version="[1.1,2.0)",
  org.eclipse.kura.db;version="[2.0,2.1)",
  org.eclipse.kura.message.store;version="[1.0,2.0)",
  org.eclipse.kura.message.store.provider;version="[1.0,1.1)",
- org.eclipse.kura.store.listener;version="1.0.0",
  org.eclipse.kura.type;version="[1.1,2.0)",
  org.eclipse.kura.util.configuration;version="[1.0,1.1)",
  org.eclipse.kura.util.jdbc;version="[1.0,2.0)",
