+++ kura/org.eclipse.kura.db.sqlite.provider/src/main/java/org/eclipse/kura/internal/db/sqlite/provider/SqliteDbServiceImpl.java
@@ -28,13 +28,13 @@ import java.util.concurrent.TimeUnit;
 import org.eclipse.kura.KuraException;
 import org.eclipse.kura.KuraStoreException;
 import org.eclipse.kura.configuration.ConfigurableComponent;
+import org.eclipse.kura.connection.listener.ConnectionListener;
 import org.eclipse.kura.crypto.CryptoService;
 import org.eclipse.kura.db.BaseDbService;
 import org.eclipse.kura.internal.db.sqlite.provider.SqliteDbServiceOptions.JournalMode;
 import org.eclipse.kura.internal.db.sqlite.provider.SqliteDbServiceOptions.Mode;
 import org.eclipse.kura.message.store.provider.MessageStore;
 import org.eclipse.kura.message.store.provider.MessageStoreProvider;
-import org.eclipse.kura.store.listener.ConnectionListener;
 import org.eclipse.kura.util.jdbc.SQLFunction;
 import org.eclipse.kura.util.store.listener.ConnectionListenerManager;
 import org.eclipse.kura.wire.WireRecord;
