+++ kura/org.eclipse.kura.util/META-INF/MANIFEST.MF
@@ -8,11 +8,11 @@ Bundle-License: Eclipse Public License v2.0
 Require-Capability: osgi.ee;filter:="(&(osgi.ee=JavaSE)(version=1.8))"
 Import-Package: org.eclipse.kura;version="[1.2,2.0)",
  org.eclipse.kura.annotation;version="[1.0,2.0)",
+ org.eclipse.kura.connection.listener;version="[1.0,2.0)",
  org.eclipse.kura.data;version="[1.1,2.0)",
  org.eclipse.kura.db;version="[2.0,3.0)",
  org.eclipse.kura.message.store;version="[1.0,2.0)",
  org.eclipse.kura.message.store.provider;version="[1.0,1.1)",
- org.eclipse.kura.store.listener;version="[1.0,2.0)",
  org.eclipse.kura.type;version="[1.1,2.0)",
  org.eclipse.kura.wire;version="[2.0,3.0)",
  org.eclipse.kura.wire.store.provider;version="1.0.0",
