+++ kura/org.eclipse.kura.wire.db.component.provider/META-INF/MANIFEST.MF
@@ -10,9 +10,9 @@ Import-Package: org.apache.logging.log4j;version="2.8.2",
  org.apache.logging.log4j.util;version="2.8.2",
  org.eclipse.kura;version="[1.2,2.0)",
  org.eclipse.kura.configuration;version="[1.1,2.0)",
+ org.eclipse.kura.connection.listener;version="[1.0,2.0)",
  org.eclipse.kura.db;version="[2.0,3.0)",
  org.eclipse.kura.message.store.provider;version="[1.0,2.0)",
- org.eclipse.kura.store.listener;version="[1.0,2.0)",
  org.eclipse.kura.type;version="[1.0,2.0)",
  org.eclipse.kura.util.collection;version="[1.0,2.0)",
  org.eclipse.kura.util.configuration;version="[1.0,2.0)",
